Use BigDecimal type for balance check before order creation #414.

+++ mm2src/coins/lp_coins.rs
@@ -261,7 +261,7 @@ pub trait MmCoin: SwapOps + MarketCoinOps + IguanaInfo + Debug + 'static {
 
     fn is_asset_chain(&self) -> bool;
 
-    fn check_i_have_enough_to_trade(&self, amount: f64, maker: bool) -> Box<Future<Item=(), Error=String> + Send>;
+    fn check_i_have_enough_to_trade(&self, amount: BigDecimal, maker: bool) -> Box<Future<Item=(), Error=String> + Send>;
 
     fn can_i_spend_other_payment(&self) -> Box<Future<Item=(), Error=String> + Send>;
